// backend/src/types/user-uestion-points.d.ts
export type UserQuestionPoints = {
  user_id: number;
  name: string;
  surname: string;
  email: string;
  scenario_id: number;
  scenario_name: string;
  question_difficulty_id: number;
  question_difficulty_name: string;
  question_id: number;
  question_text: string;
  selected_option_points: number;
  timestamp: string;
};